\name{add_cell_info}
\alias{add_cell_info}
\title{
add annotation on each cell  


}
\description{
add annotation on each cell  


}
\usage{
add_cell_info()
}
\details{
add the names and the index of the track in each cell. This function is only for demonstration purpose.  


}
\value{
no value returned  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
\examples{
initialize_layout()
add_cell_info()
}
