\name{gtrellis_info}
\alias{gtrellis_info}
\title{
Add annotation on each cell  


}
\description{
Add annotation on each cell  


}
\usage{
gtrellis_info()
}
\details{
The function adds name and index for each cell in the corresponding track.  It is only for demonstration purpose.  


}
\value{
No value is returned.  


}
\author{
Zuguang Gu <z.gu@dkfz.de>  


}
\examples{
gtrellis_layout()
gtrellis_info()
}
