\name{gtrellis_show_index}
\alias{gtrellis_show_index}
\title{
Show index on each cell
}
\description{
Show index on each cell
}
\usage{
gtrellis_show_index()
}
\details{
The function adds name and index of track for each cell. 
It is only for demonstration purpose.
}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
gtrellis_layout()
gtrellis_show_index()

}
