\name{add_ideogram_track}
\alias{add_ideogram_track}
\title{
  add ideogram track  


}
\description{
  add ideogram track  


}
\usage{
add_ideogram_track(species = NULL, i_track = get_current_cell_meta_data("i_track")+1)
}
\arguments{
  \item{species}{species}
  \item{i_track}{which track}

}
\details{
  add an ideogram track 


}
